fix: translations and use container component
