+++ src/components/ReserveSubheader.tsx
@@ -1,3 +1,4 @@
+import { Trans } from '@lingui/macro';
 import { Typography } from '@mui/material';
 import Box from '@mui/material/Box';
 import React from 'react';
@@ -21,7 +22,7 @@ export function ReserveSubheader({ value, rightAlign }: ReserveSubheaderProps) {
     >
       {value === 'Disabled' ? (
         <Typography component="span" sx={{ mr: 0.5 }} variant="secondary12" color="text.muted">
-          (Disabled)
+          (<Trans>Disabled</Trans>)
         </Typography>
       ) : (
         <FormattedNumber
